Sequential Resampling Random Walks
In order to be able to work on an extension of his research, my research mentor, Dr. Kevin Liu, had me write my own code for an algorithm his lab had come up with earlier for DNA sequence re-sampling. The idea behind the algorithm is to increase statistical support estimation by making more accurate re-samples. In other sets of data, who your neighbor is might not matter, but when it comes to genetics, base pair neighbors are important, therefore the re-sampling procedure for support estimation should take this into account.