Budget Breakdown Program
This was a project I did as an honors option for my Object-Orientated Software Design class. It is a simple budgeting program with a GUI that allows a user to input transactions, and view the display breakdown of different months. It can also save all current information into an XML file that a user can load up into the program later.